utilsディレクトリにcalcProgressメソッドを切り出し
